wctc4xxp: Cleanup in-flight commands when halting due to hardware error.
On one system I was seeing the board reset in the middle of a
transaction. Any commands that were on the response list when this would
happen would never be completed and the process would then be stuck in
an uninterruptible sleep. This change also prevents the driver from
sleeping in timer context, which would result in a kernel panic.

This change at least lets an error message propogate back to the user.

DAHDI-880

Signed-off-by: Shaun Ruffell <sruffell@digium.com>

Origin: http://svnview.digium.com/svn/dahdi?view=rev&rev=10082

Conflicts:

	drivers/dahdi/wctc4xxp/base.c

git-svn-id: http://svn.asterisk.org/svn/dahdi/linux/branches/2.4@10084 a0bf4364-ded3-4de4-8d8a-66a801d63aff